PIC-405 Provider API to allow flush of memory redo-log into storage (#204)
* Internal API to flush redo-log to storage
* Test code showing the functionality of reading contents from storage

+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/server/space/redolog/storage/SqliteStorageLayer.java
@@ -54,7 +54,7 @@ public abstract class SqliteStorageLayer<T extends IReplicationOrderedPacket> {
             if (!path.toFile().mkdirs()) {
                 throw new StorageException("failed to mkdir " + path);
             }
-        } else {
+        } else if (!config.shouldUseExistingDBFileOnStart()){
             deleteDataFile();
         }
 
@@ -67,7 +67,9 @@ public abstract class SqliteStorageLayer<T extends IReplicationOrderedPacket> {
             logger.error("Failed to initialize Sqlite RDBMS", e);
             throw new StorageException("failed to initialize internal sqlite RDBMS", e);
         }
-        createTable();
+        if (!config.shouldUseExistingDBFileOnStart()){
+            createTable();
+        }
     }
 
     private Connection connectToDB(String jdbcDriver, String dbUrl, String user, String password, SQLiteConfig sqLiteConfig) throws ClassNotFoundException, SQLException {
